Split sanity code into separate file.
